Put JSON-RPC types in a separate package
