+++ kore-rpc-types/kore-rpc-types.cabal
@@ -1,0 +1,200 @@
+cabal-version: 2.2
+name:           kore-rpc-types
+version:        0.60.0.0
+description:    Please see the [README](README.md) file.
+category:       Language
+homepage:       https://github.com/runtimeverification/haskell-backend#readme
+bug-reports:    https://github.com/runtimeverification/haskell-backend/issues
+author:         Runtime Verification Inc
+maintainer:     ana.pantilie@runtimeverification.com
+copyright:      2018-2021 Runtime Verification Inc
+license:        BSD-3-Clause
+build-type:     Simple
+
+source-repository head
+  type: git
+  location: https://github.com/runtimeverification/haskell-backend
+
+flag threaded
+  description: Use the threaded runtime. Recommended to disable for profiling.
+  manual: True
+  default: True
+
+common haskell
+  -- Common Haskell dialect
+  default-extensions:
+    BangPatterns
+    -- GHC2021 prep
+    BinaryLiterals
+    -- end
+    ConstraintKinds
+    DataKinds
+    DefaultSignatures
+    DeriveAnyClass
+    DeriveDataTypeable
+    DeriveFoldable
+    DeriveFunctor
+    DeriveGeneric
+    -- GHC2021 prep
+    DeriveLift
+    -- end
+    DeriveTraversable
+    DerivingVia
+    DuplicateRecordFields
+    EmptyCase
+    -- GHC2021 prep
+    EmptyDataDeriving
+    -- end
+    ExistentialQuantification
+    FlexibleContexts
+    FlexibleInstances
+    FunctionalDependencies
+    GADTs
+    GeneralizedNewtypeDeriving
+    -- GHC2021 prep
+    HexFloatLiterals
+    -- end
+    InstanceSigs
+    -- GHC2021 prep
+    ImportQualifiedPost
+    -- end
+    KindSignatures
+    LambdaCase
+    MultiParamTypeClasses
+    NamedFieldPuns
+    -- GHC2021 prep
+    NamedWildCards
+    NumericUnderscores
+    -- end
+    OverloadedStrings
+    PatternSynonyms
+    -- GHC2021 prep
+    PolyKinds
+    PostfixOperators
+    -- end
+    RankNTypes
+    ScopedTypeVariables
+    StandaloneDeriving
+    -- GHC2021 prep
+    StandaloneKindSignatures
+    -- end
+    Strict
+    -- GHC2021 prep
+    TupleSections
+    -- end
+    TypeApplications
+    TypeFamilies
+    TypeOperators
+    TypeSynonymInstances
+    UndecidableInstances
+    ViewPatterns
+  default-language: Haskell2010
+  build-depends: base >=4.7
+  ghc-options:
+    -fno-ignore-asserts
+    -Wall
+    -Wcompat
+    -Widentities
+    -Wincomplete-uni-patterns
+    -Wincomplete-record-updates
+    -Wredundant-constraints
+    -Wnoncanonical-monad-instances
+    -Wmissing-deriving-strategies
+  ghc-options:
+    -j6
+  ghc-options:
+  if impl(ghc >= 8.4)
+    ghc-options: -Wmissing-export-lists -Wpartial-fields
+  if impl(ghc >= 8.8)
+    ghc-options: -fwrite-ide-info -hiedir=.hie
+
+common library
+  -- Dependencies of the library, shared in common with the tests
+  build-depends: aeson >=1.4
+  build-depends: aeson-pretty == 0.8.9
+  -- build-depends: array
+  -- build-depends: async >=2.2
+  -- build-depends: binary >=0.8.8.0
+  -- build-depends: bytestring >=0.10
+  -- build-depends: clock >=0.8
+  -- build-depends: co-log >=0.4
+  -- build-depends: comonad >=5.0
+  -- build-depends: compact >= 0.2.0.0
+  -- build-depends: conduit >= 1.3
+  -- build-depends: conduit-extra >=1.3
+  -- build-depends: containers >=0.5.8
+  -- build-depends: cryptonite >=0.25
+  -- build-depends: data-default >=0.7
+  -- build-depends: decision-diagrams >=0.2.0.0
+  -- build-depends: deepseq >=1.4
+  build-depends: deriving-aeson >=0.2
+  -- build-depends: directory >=1.3
+  -- build-depends: distributive >=0.6
+  -- build-depends: errors >=2.3
+  -- build-depends: exceptions >=0.10
+  -- build-depends: extra >=1.6
+  -- build-depends: fgl >=5.7
+  -- build-depends: filepath >=1.4
+  -- build-depends: free >=5.1
+  -- build-depends: generic-lens >=1.1
+  -- build-depends: generics-sop >=0.5
+  -- build-depends: ghc-events
+  -- build-depends: ghc-trace-events >=0.1
+  -- build-depends: gitrev >=1.3
+  -- build-depends: graphviz >=2999.20
+  -- -- Pin hashable to make sure cabal build uses the one Nix & Stack builds do
+  -- build-depends: hashable ==1.4.2.0
+  -- build-depends: haskeline ==0.8.*
+  -- build-depends: ghc-bignum >=1.2
+  build-depends: json-rpc >=1.0
+  -- build-depends: lens >=4.17
+  -- build-depends: lifted-async
+  -- build-depends: lifted-base
+  -- build-depends: logict >=0.7
+  -- build-depends: matrix >=0.3.6.1
+  -- build-depends: megaparsec >=7.0.4
+  -- build-depends: memory >=0.15
+  -- build-depends: mmorph >=1.1
+  -- build-depends: monad-control
+  -- build-depends: monad-logger >=0.3.36
+  -- build-depends: monad-validate
+  -- build-depends: mono-traversable >=1.0
+  -- build-depends: mtl >=2.2
+  -- build-depends: multiset >=0.3.4.3
+  -- build-depends: optparse-applicative >=0.14
+  -- build-depends: parser-combinators >=1.1
+  -- build-depends: pqueue >=1.4.1.3
+  build-depends: prettyprinter >=1.2
+  -- build-depends: process >=1.6
+  -- build-depends: profunctors >=5.3
+  -- build-depends: recursion-schemes >=5.1
+  -- build-depends: semialign >=1
+  -- build-depends: sqlite-simple >=0.4
+  -- build-depends: stm >=2.5
+  -- build-depends: stm-conduit >= 4.0
+  -- build-depends: streams
+  -- build-depends: tar >=0.5
+  -- build-depends: template-haskell >=2.14
+  -- build-depends: temporary >=1.3
+  build-depends: text >=1.2
+  -- build-depends: these >=1.0
+  -- build-depends: time >=1.9.3
+  -- build-depends: transformers >=0.4
+  -- build-depends: transformers-base
+  -- build-depends: unliftio >= 0.2
+  -- build-depends: unordered-containers >=0.2
+  -- build-depends: vector >=0.12
+  -- build-depends: witherable ==0.4.*
+  -- build-depends: yaml >=0.11.5.0
+  -- build-depends: zlib >=0.6
+  -- build-tool-depends: happy:happy
+  -- build-tool-depends: alex:alex
+
+library
+  import: haskell
+  import: library
+  exposed-modules:
+    Kore.JsonRpc.Types,
+    Kore.Syntax.Json.Types
+  hs-source-dirs:
+    src
