+++ stack.yaml
@@ -43,6 +43,7 @@ nix:
 # will not be run. This is useful for tweaking upstream packages.
 packages:
   - kore
+  - kore-rpc-types
 
 # Dependency packages to be pulled from upstream that are not in the resolver
 # (e.g., acme-missiles-0.3)
